improve logging of errors in simlpestreams
We were throwing away extremely informative error messages, mostly because
of a missing ! in an if statement.  We don't need to log that there aren't
any matching tools, but we should be logging things like json unmarshaling
errors... the currernt code was doing the opposite of this.  Also added
one more trace error logging in a place that seems to be useful.