Expose max-log-size and max-log-age as controller configuration.
This commit adds the max-log-size and max-log-age config
directives in the controller level.

Both config directives were previously declared as constant
variables defined on github.com/juju/juju/workers/dblogpruner.go.

The rationale for exposing those configuration options comes
by the fact that some users needs to adjust the logs retention time
to a value that is better suitable for the environment.

This commit fixes LP: #1566545

Signed-off-by: Jorge Niedbalski <jnr@metaklass.org>