worker/migrationminion: implement VALIDATION phase
The migrationminion now responds to the VALIDATION phase. It will
attempt to connect to the target controller and run an external
validation function against it. The success or failure of these checks
are reported using the pre-existing minion reporting API.

The migration will be aborted if any minion reports that it wasn't
able to validate against the target controller. This gives a much
higher degree of confidence that the migration is going to succeed.