state: don't force-destroy manual machines
When destroying a model, we were force-destroying
all non-manager machines. We should not force-destroy
manual machines, as doing so may lead to applications
being left running on the machine. Instead, just
issue a non-force destroy. If the machine is in a
bad state, the user will have to force-destroy
via "juju remove-machine --force".

Fixes https://bugs.launchpad.net/juju-core/+bug/1475212