provider/azure: speed up instance registration
Don't wait for VM or VM extensions to be be
fully provisioned before returning from StartInstances.
We intercept the responses and rewrite the provisioning
state from "Creating" (for example) to "Succeeded".
This stops the Azure SDK from synchronising the
request.

There is a regression in functionality introduced by
this: if VM provisioning fails asynchronously, then
it will not be possible to identify the cause of the
failure without looking in the Azure console. This
will be rectified in a follow-up.

Fixes https://bugs.launchpad.net/juju-core/+bug/1614689