Add pingerFacadeVersion field for Ping call
Previously the Ping method called BestFacadeVersion each time it was
called. Inefficient as the version isn't going to change during the
life of the connection. Worse it creates a race in tests where the
underlying facadeVersions map could be deleted and recreated in
setLoginResult. This doesn't happen in normal usage so this fixes the
race by adding a pingFacadeVersion field to the state struct and
initializing it simultaneously.

Refs: https://bugs.launchpad.net/juju-core/+bug/1614732