Fixes lp:1596853 & refactoring deploy cmd for UT
This patch accomplishes a few things:

* Reduce the number of connections to the API server
  During deploy, we were opening several clients and repeatedly opening/closing new API connections. We now open one connection at the root of the command in `Run(...)` and pass it through as needed. In my tests, the connections a deploy of a local charm requires to work has been reduced from 4 to 1, and as a result, deploys will be fater -- especially in scenarios where the connection to the controller is slow.

* The deploy command now takes in an API
  This allows for unit tests to be written against this command.

* The API passed in has begun to be pared down from api.Connection.
  This work is not complete, but it provides a strong start to build on.

Some tests required touch-ups to work with the new API being passed in. Some code could be deleted as it was providing obviated abstraction around opening new API connections.

Please note that this only represents the start of a larger body of work. Currently, the way of mocking out some of the tests relies too much on knowledge of implementation. As the API interface being passed in shrinks, this problem will shrink as well.