cmd/jujud/agent/model: Stop engine if model is removed
When the apicaller encounters a "model not found" error convert it to
ErrRemoved so that the dependency engine for the model is
stopped.

This fixes a race that occurs during model migration or model removal
where the modelworkermanager's watcher fires as the model is removed
causing the model's engine - which had just stopped - to be
restarted. The engine is then stuck in zombie state because an API
connection can't be established in order to find out that the model is
dead. This left the model's dependency engine running with the
apicaller continually trying - and failing - to connect to the
long-gone model.

Fixes LP 1614809