provider/azure: reduce API calls in provisioning
Reduce the number of Azure API requests we make
when provisioning an instance. Instead of creating
and then fetching resources, we craft the resource
IDs ourselves and use them when constructing cross-
resource references.

A significant change here is that controller machines
and non-controller machines are now allocated to
separate subnets. This means that we can create a
single network security group rule to allow API
access for all controller machines. The subnets are
on the same virtual network, and the machines can
still communicate with each other across them.