add-model to take cloud/region positional arg
The add-model command now has an optional positional
argument, which can be the name of a cloud, a region,
or both (cloud/region). If it is a cloud, then the
cloud's default region will be used, if there are
any regions. If it is not a cloud, we check if it is
a region in the default cloud.