provider/maas: bridge all interfaces
Ensure all interfaces get bridged. We would previously not bridge
interfaces that were deemed inactive. An active interface was previously
one which had either a static address or was using DHCP. This contraint
does not work for DTAG setups because an interface that wasn't
active (or bridged) would not be part of the interface set in a
container. These (previously) inactive interfaces may actually be
required for Neutron which will want to use the underlying raw
device (at L2), as opposed to requiring a L3 address. This commit
ensures all interfaces get bridged and, as such, those interfaces will
be available in containers.

Partially fixes [LP:#1566791](https://bugs.launchpad.net/juju/+bug/1566791)