cmd/modelcmd: provide flag to suppress browser login
This change makes logging in through a web browser the default,
as that works even if the user hasn't logged in previously,
but they can still get the current behaviour by using the -B
flag.

This requires all SetFlags implementations to call the SetFlags
implementation of their embedded Base object (which is only
good practice anyway) and we also simplify some of the resource
command logic to use the shared APIContext implementation
so that gets the benefit too, and remove some of the more
unnecessary-seeming abstractions in that area.