Take the introspection worker out of the agent engine
The introspection worker cannot and should not be in the engine
itself as it reports on the engine, and other aspects of the
runtime. If we put it in the engine, then it is mostly likely
shut down in the times we need it most, which is when the agent
is having problems shutting down.