Fixes lp:1596858 and refactors the deploy command
This work builds on previous work to refactor the deploy command to take
in an API interface instead of instantiating clients throughout. This
brings us closer to achieving that goal and allows us to write unit
tests for the majority of the deploy command (majority of the current
tests are so-called "full-stack").

The highlights of this patch are:

- Modifies the output when deploying bundles to the format suggested in
  the bug. The details can still be achieved by passing the `--debug`
  flag to the command.

- Removed redundant checks for format of output when deploying
  bundles. This check is done in a single test and need not be
  replicated everywhere.

- Introduced a type, `callMocker` to assist with mocking interfaces for
  tets. This can be found in deploy_test.go. In the long run, it would
  be nice for this type to move over to juju/testing to live with the
  `Stub` class.