Overhaul local login macaroon auth
This is a long overdue overhaul of the local
login macaroon authentication in Juju, bringing
macaroon authentication for local users in line
with how we do it for external users.

The controller now acts as a discharger for
macaroons authenticating local users. When you
first attempt to connect without a password you
will be handed a macaroon with the location
pointing back at the controller. The client
has a special "web-page" visitor which can
interact with the controller to obtain a caveat
discharge macaroon with which it can authenticate.

The macaroon field in accounts.yaml has been
dropped, in favour of storing macaroons only in
the cookie jar ($JUJU_COOKIEFILE or ~/.go-cookies).

There is currently no browser-based login page
for local users, though this could be added later.

Fixes https://bugs.launchpad.net/juju/+bug/1603176