Refactor currentKeyDataForDelete so order is kept
Because the ssh key data from disk was being stored in a map then
reserialised, the order of the keys is not preserved. This could
be a cause of minor confusion, but also meant the new test for
key deletion would sometimes fail.

This change refactors the helper so the returned slice is the key
store, and the maps are just a way to do faster lookups.