migration: Check for active migration in target
This check is necessary because there is a window between the REAP phase
and then end of the DONE phase where a model's documents have been
deleted but the migration isn't quite done yet. Migrating a model back
into the controller during this window can cause the migrationmaster
worker to be stopped and other workers to be incorrectly restarted.

This fixes https://lpad.tv/1611391