provider/azure: use ARM templates
This diff fundamentally changes how we provision
resources in Azure. Instead of creating the resources
individually, we now create them using ARM templates.
Templates describe all of the resources we require,
and their dependencies; and Azure parallelises the
creation of those as well as it can. In addition, we
now delete VMs in parallel.

The main benefit is that bootstrap time is back to how
it was before we updated the Azure SDK recently. There
are several other improvements:
 - few API requests
 - VM deployments are now fully parallelisable
 - add-model is faster, as we now only create the
   resource group synchronously
 - simpler code and tests
 - listing instances is now less awkward; each instance
   has a corresponding deployment. If that deployment
   exists, the instance exists. Compare and contrast to
   how it was before, where an instance's existence was
   based on whether its *NIC* existed.

There are a couple of major changes that are required
to enable this:
 - we no longer generate a random storage account name,
   and cycle through until we choose one. Instead, we
   take the last 10 bytes of the model UUID, format it
   as hex, and prepend "juju". This should be enough to
   ensure a unique name. If it is not, we will need to
   create the storage account at the same time as the
   resource group, which will slow down deployments.
 - we no longer make an API query to determine the next
   available private IP address. Instead, we compute
   the address based on the machine ID. One problem this
   introduces is that if we cycle through ~4K machine
   IDs in a model, then we won't be able to create any
   new instances. We can address this later by maintaining
   in memory the set of in-use IP addresses.