worker/migrationmaster: Support external control
If a migration is marked as externally controlled, the migrationmaster
will simply wait for the migration to end without performing any of its
normal actions.

This change also features some major refactoring of the worker
tests. Large amount of duplicated code and boilerplate has been removed,
making the tests easier to read and maintain.