state: Charm now has Life
...along with the associated Destroy, Remove, and Refresh methods.
Only implemented for local charms; store charms are immortal, as
discussed in comments.

All attempts to reference a charm now check the charm's life; the
charm's life will not advance when any reference exists. Once a charm
has been destroyed, it can be removed; at which point any storage it is
using will be reclaimed.