Better error message for bad MAAS endpoint URL
Fixes http://pad.lv/1623184

When the MAAS endpoint is wrong (in the bug's case, didn't end in MAAS)
Juju would report "ERROR MAAS version 1.9 or more recent is required not
supported". Change that error to indicate that actually we couldn't work
out what version of MAAS we were talking to, and the endpoint is
probably wrong.

Fix the other badly-worded NotSupported error as well.

Drive-by fix for the environ tests that created a TestMAASObject and
then never used it.