provider/azure: update auth-types
We introduces two new auth-types, and deprecate
use of "userpass", for Azure. The userpass auth-type
is superseded by service-principal-secret, which is
identical except that it does not include the
tenant-id field. There is a new "interactive"
auth-type which is defined, but not yet supported.

If a user uses "userpass", they will receive a
warning explaining that it is deprecated, and what
they need to do to migrate. The credential will
continue to work, we'll just ignore the tenant-id.