provider/azure: add interactive auth-type
This diff adds the interactive auth-type
to Azure. When adding a credential with
the "interactive" auth-type, the user is
prompted for their subscription ID; Juju
will then initiate device-code auth, which
involves the user opening up a URL and
copying in a code. Juju then waits for a
token from the Microsoft login server,
and uses it to set up a service principal
automatically.