provider/azure: update AD client
This updates the Azure Active Directory client, in preparation for a
follow-up which adds the interactive auth-type to the Azure provider.
We do not need to manipulate Applications, only Service Principals;
so the applications client has been removed. The users client has been
added so that we can log the identity of the user that authenticated.

The WithOdataErrorUnlessStatusCode function is a copy of
WithErrorUnlessStatusCode from the autorest/azure package, except
that it expects the error object's name to be "odata.error" instead
of just "error", and does not extract the error message. This is not
yet used, and will be tested in a follow-up.