provider/gce: don't set sshKeys instance metadata
The GCE provider is bailing from StartInstances if
authorized-keys is empty. It should not be doing
this.

Also, we're only looking at authorized-keys in the
GCE provider to set the sshKeys instance metadata.
This is what GCE uses to keep track of the SSH keys
that *it* manages on an instance. We should not be
setting this; using cloud-init is sufficient.

Fixes https://bugs.launchpad.net/bugs/1625881