upgrades: Remove old apiserver charm get cache
This takes care of any charm get cache created by the apiserver in an
RC1 or earlier release.

Note the new style of testing upgrade steps (using the new findSteps
helper). This approach both ensures that the step is correctly wired up
and does what it's supposed to in the one test, and doesn't require
much in export_test.go.

The old test helpers have been removed.