state: update storage constraints in SetCharm
Application.SetCharm now supports updating
storage constraints at the same time as
updating an application's charm, and so we
relax the rules around changes to a charm's
stores.

If a new required store is added, we will add
max(min-count, constraint-count) instances
to each unit, where min-count is the minimum
count specified in the charm metadata, and
constraint-count is the user-specified count,
if any. If constraint-count < min-count, we
will return an error explaining that the
constraints do not satisfy the requirements.

The minimum count for existing stores may
also be changed now. If any unit does not
already have enough storage instances to
satisfy this requirement, new instances will
be added so long as the storage constraints
allow it.

**QA**

(With https://github.com/juju/juju/pull/6312)

1. bootstrap and deploy a charm with a
   single optional store, with count=1
2. modify the charm to make said store
   required, with a minimum count of 2;
   and add another store with a minimum
   count of 3
3. upgrade-charm to said charm, without
   specifying --storage. Observe that it
   fails because the existing store's
   constraints (count=1) do not satisfy
   the new minimum (count>=2).
4. upgrade-charm to said charm, with
   --storage existing-store=2, but without
   specifying anything for the new store.
   This will succeed; there will be an
   additional storage instance for the
   existing store, and three instances
   for the new store with the default
   storage constraints.
5. Do-over, but this time with --storage
   for the new store also, specifying
   new-storage=4. Observe that the user-
   specified constraints are used for
   the new store, and four instances are
   created.