apiserver: add certificate changing test
Currently the certificate changing behaviour seems to be
untested in the apiserver package.

We also make the cert change listener a little more
like the other worker goroutines started by the
API server - it now listens to the API server tomb
and is waited for when the API server stops.

Also we move the changeCertListener code towards end of file
so it doesn't clutter the higher level code.