apiserver: implement LetsEncrypt support
This enables a controller to be contacted via a real
domain name - it will obtain an official HTTPS certificate
from the letsencrypt service.

It's not easy to test this without doing it for real, so to QA:

- bootstrap a controller with api-port=443
- set the IP address associated with a domain you control to the public
IP address of the new controller.
- edit controllers.yaml to change the addresses associated
with the controller to <yourdomain>:443.
- use the controller.

Note that as is, this is open to the DoS vulnerability described
in https://godoc.org/golang.org/x/crypto/acme/autocert#Manager
under the HostPolicy field. We really want to whitelist a single
domain, but that involves adding controller configuration, so we
reserve that for a future PR.