Corrects misleading error message when deploying
This addresses lp:1616200, with caveats.

Previously deploying a charm with an incorrect name would:
- Add the charm to the environment
- Display:
> "ERROR bad charm URL in response: URL has invalid charm or bundle name: "local:<series>/<charm>-<rev>""

Subsequent deploys of the charm with an incorrect name would display:
> ERROR POST https://10.1.118.174:17070/model/ba258aba-b24d-497f-8a38-aebba03105ee/charms?revision=0&schema=local&series=xenial: bad charm URL in response: URL has invalid charm or bundle name: "local:<series>/<charm>-<rev>""

This patch corrects the chain of errors to have proper causes which implicitly corrects the error message to:
> cannot upload charm: name "b_1605096" not valid

or series, or whatever the case may be. However, because it fails faster, it takes a different path through the code. The bad charm won't be added to the `charms` mongo collection (good), but we now always see the verbose POST prepended in front of the error (bad).

To correct this properly, I think we should modify the juju/httprequest `Do` method. I plan on doing this in a subsequent patch.