all: use Go standard spelling for acronyms
The code base is currently inconsistent about whether to use
"Api" or "API", or "Numa" or "NUMA", etc.

This PR changes the code to use the standard Go convention of
capitalising the whole acronym when the first letter is capital.
We've done this for API, NUMA, HTTPS, HTTP, FTP and URL.

It's almost entirely mechanical (the only manual change was in
cmd/modelcmd/base.go where there was an instance variable
named modelApi but a method named modelAPI.

There are some external packages that could also benefit
from this change (notably github.com/juju/utils/proxy).