Move the GetBundleChanges call in a new Bundle facade.
Also make the new facade available on both controller and model connections.

This is done so that GUI users can render the bundle uncommitted state even without a model connection.
Note that, even if this is a backward incompatible API change, I think we can be quite sure that the GUI is the only client of this API.