Start the migratation of state tests to a testing clock
Continue to remove time.Now() in favour of ZeroTime() and NonZeroTime()
from the testing package in state tests, and begin the required
migration of clocks used in state tests to testing clocks with
NonZeroTime(), not the wallclock.

StateSuite and ConnSuite now each have a 'wallclock' partner, used in
tests that require the wallclock until they are migrated to the testing
clock.