openstack: defer querying volume endpoints
We were attempting to identify the volume
endpoint for an OpenStack environment when
the Environ was opened. This doesn't work,
as the client is not authenticated at that
time, and so the endpoints aren't known.

We could authenticate at environ-opening
time, but that would slow down prechecking,
among other things. Instead, we change the
StorageProviderTypes method to return an
error, and query the endpoint there.

Fixes https://bugs.launchpad.net/juju/+bug/1615095