state/workers: dynamic lease managers
If a lease (leadership or singular) manager worker
is restarted because of a Mongo timeout, then the
singular API server worker will not currently
recover. This is because the State.SingularWorker()
(and similar Leadership methods) return only the
original worker; the worker returned will continue
to fail until a new worker is acquired through
State.SingularWorker.

Instead of forcing the callers to do that, we
instead wrap the returned lease manager so that
the individual methods operate on the active
manager.

Fixes https://bugs.launchpad.net/juju/+bug/1628206