provider/lxd: clean up client certificates
We now record the fingerprint of the client
certificate in the instance metadata of
controller machines. When destroying the
machine, we check the metadata for a
certificate fingerprint, and then remove
the certificate with that fingerprint from
the server.

Because we did not previously record the
certificate fingerprint, this changes only
enables us to remove certificates for newly
created controllers.

Fixes https://bugs.launchpad.net/juju/+bug/1616346