provider/ec2: add support for ap-south-1 (Mumbai)
This diff adds support for the new(ish) ap-south-1
region in Mumbai. Rather than adding to the various
hard-codings we have about the place, the ec2
provider is updated to honour the values in
clouds.yaml.

There are numerous changes to enable this:
 - the default public clouds have been updated to
   add ap-south-1. The existing endpoints were all
   wrong, and have been updated.
 - the hard-coded instance types are removed from
   provider/ec2/instancetypes.go; this is replaced
   with generated data using the Price List API.
   This is still static data, and needs to be
   updated periodically. We may later update the
   provider to periodically pull this information
   down from within the controller.

The S3 code is not being used anymore, and so has
been deleted.

Fixes https://bugs.launchpad.net/juju/+bug/1612645

Requires https://github.com/juju/juju/pull/6374