mongo: Use a timeout when dialing
When dialling a mongodb server, use the timeout from the DialOpts. This
helps mgo to respond faster when a mongodb node goes down, and assists
with HA failover.

When not using a custom dial function, mgo will use the timeout with
net.DialTimeout in the same way.

Also switched to passing the dial function using the DialServer field of
mgo.DialInfo instead of the deprecated Dial field.

Some logging has been tidied up too: refer to mongodb instead of mongo
and adjusted log levels.

Part of the fix for https://bugs.launchpad.net/juju/+bug/1588224