provider/ec2: update default controller instance
In a VPC account, we'll create a t2.medium for
controller instances by default now. Non-controller
instances are unchanged for now, and remain m3.medium
by default. Controller instance root disk now defaults
to 32GiB, but can be overridden with constraints.

There is a fix to the ec2instancetypes generator:
we were expecting "currentGeneration" to be present
in the attributes always, but it is not there for
some. We assume they are current if it is missing,
which is the case with the ones that have it missing
currently.

Fixes https://bugs.launchpad.net/juju/+bug/1535838