Add StatePool.Remove to close States
This will be called when the underlying model is removed to prevent
State instances from leaking. Since State instances can be shared
between requests (in the API server for example), this requires keeping
a count of the times a state has been requested from the pool with
Get. Also add StatePool.Put so clients can indicate when they've
finished with a State.

If Remove is called for a State and there are no outstanding references,
it will be closed and removed immediately. If there are outstanding
references, it's marked for removal and removed and closed when the last
use is Put back.

At the moment this relies on the client to watch for model removals and
remove States from the pool, which is not ideal. There are two places
using StatePool and they use different types for managing lifetimes
unfortunately - the API server uses a tomb (v1), and the AllModelWatcher
uses a catacomb. It seemed like I'd have to convert the API server to
use a catacomb to be able to do the watching inside the StatePool, so I
did this instead.

Part of fixing http://pad.lv/1625774