lxdclient: add a flag indicating whether to check the local bridge config
The problem here is a chicken and egg problem: the container type needs to
connect to the local LXD to figure out how to configure the bridge,
but lxdclient has code to check the (old style) bridge configuration when
connecting to the local LXD.

Really, this code that checks the bridge configuration should live in the
LXD provider, but I suspect that there'd need to be some abstraction
untangling to do that.

Instead, for now, let's add a boolean flag to Connect() that just indicates
whether or not we should verify the bridge configuration.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>