api: add SNIHostName to Info
This means that we can have an arbitrary API addresses
in api.Info, including resolved IP addresses, but still connect
using the officially signed certificate to check. If there's
a private Juju CA cert available, we use that by preference
to make it possible to connect even if the server cannot
obtain an officially signed certificate.

We also add a DialWebsocket field to api.DialOpts
so that we can use a fake dialer function in tests.