Remove Ping from api.Connection interface
As per pre-existing comments, there should only be one way to check
whether an API connection is alive. Given that connections now do a
great job of checking themselves - the results of which are exposed via
Broken - the Ping method was removed.

This had some bigger-than-expected fallout in the agent code but this
can be cleaned up by making a similar change to State (later).