controller: allow HTTP identity URL if provided with public key
The reason for the restriction of identity-url to HTTPS only
is that we don't want to fetch its public key insecurely.
This PR allows an HTTP URL if the public key is provided (matching
the check made in the bakery code), and also adds a few validation tests
that were inexplicably missing from the controller package.