apiserver: Separate clock for pinger
Allow a separate clock to be used for the apiserver's connection ping
timeout functionality. Without this it's impossible to test the ping
functionality reliabily as there are a number of goroutines in the
apiserver which wait on the server's clock.

If a pinger clock isn't specified the server's main clock is used.

Likely fix for https://bugs.launchpad.net/juju/+bug/1632412