apiserver/observer: Reduce API conn log levels
The start and end of every API connection was being logged at INFO. This
is not generally useful to the average user and is strange when view
logs at INFO because all you see is "API connection from ... ", followed
by "API connection terminated..." with nothing in between.

These messages have been lowered to DEBUG so that they're at the same
level as the other API handling log output.