Fix TestCertificateDNSUpdatedInvalidPrivateKey
This test contained an easy to reproduce race. The way it was waiting
for the certificate updates meant that it would often end up seeing the
certificate before the update was happened causing certificate parsing
to fail.

The test has been rearranged to eliminate the race and the polling for
changes. The certificate in the agent's config is now only checked once
it is known that agent startup has occurred.