apiserver: listen on localhost in tests
The old newServer code listened on ":0" not localhost,
but that wasn't a problem because nothing relied
on dialing the address taken from the listener.

When the two server-starting pieces of code were consolidated
into one, we didn't change the code to listen on "localhost:0"
instead of ":0", which causes the tests to fail on Windows
where dialing 0.0.0.0 doesn't count as dialing localhost.

Fixes https://bugs.launchpad.net/juju/+bug/1466011.