Remove deprecated streams directory
This addresses lp:1613858.

When running `juju metadata generate-tools`, Juju would attempt to look in a deprecated path "releases" for tools to generate streams. If you specified a value in the `--stream` flag (e.g. "devel"), Juju would work correctly. This was because we set this flag to the 1.x default of "releases" (notice: not "released").

This patch sets the default of the `stream` flag to envtools.ReleasedStream to give some transparancy to the user and follow a more common idiom. This also fixes the bug.

I also added some `errors.Trace` around error paths in the functions I touched.