Fix LXD domain name to be ".lxd" to match the default network configuration
This is a temporary fix until we can discover the correct domain.

Don't set manage_etc_hosts in cloudconfig/containerinit because it isn't needed for non-LXD provider cases (or so it seems).

Tested by checking that .lxd is the domain used in LXD environments
1. Bootstrap a lxd environment + add a machine
2. juju ssh 0 hostname -f
3. ...should show hostname.lxd

Tested by checking that .lxd is not used in non-LXD environments (I tried MAAS)
1. Bootstrap a non-LXD environment + add a machine
2. juju ssh 0 hostname -f
3. ...should show $hostname.$something where $something != 'lxd'.

I have also tested by deploying a rabbitmq-server cluster on LXD, which needs name resolution to work.

Fixes:
http://pad.lv/1563271
http://pad.lv/1633126