worker/logforwarder: Make tests reliable
The tests were using a convoluted mess of channels in an attempt to
synchronise the multiple goroutines being tested with the test
code. This was hard to understand and work with and is likely to be
caused the observed intermittent test failures.

Changes:
- use independent stubs for the stream and sender to make it easier to
  assert calls and to have faked errors always be returned by the
  correct instance.
- use a single simpler "activity" channel in the stub sender to
  synchronise test and code-under-test progress
- make checks more reliable by only checking stubs once worker has shut
  down.
- removed unnecessary record sending goroutine in stub stream (just
  stack a buffered channel instead.

Likely fix for https://bugs.launchpad.net/juju/+bug/1606568